Add transition on keyguard showing
This transition helps with ensuring the correct wallpapers when
keyguard is showing.
We are experiencing issues with the wallpapers changing visibility
during a transition, thus having an incorrect finish state: a wallpaper
can not become immediately invisible while in a transition (because it
is expected to commit invisibility only at the end), which
implies that if the wallpaper is collected while becoming visible, the
requests to become invisible during the transition period get ignored (because at the end of the transition we are expected to not make it invisible, since it just became visible at the start).
This implies that locking keyguard while there is an ongoing transition
leads to the wrong wallpaper showing on lockscreen, because we can not
hide the home wp simply by requesting to adjustWallpapers. Therefore, we
create a new transition to avoid handling unexpected visibility changes
on ungoing transitions.

Bug: 327970608
Bug: 323226308

Test: manual - delay execution of
KeyguardViewMediator#updateActivityLockscreenState, wait for keyguard to
unlock and observe lockscreen
Test: manual - unlock keyguard, open app, lock keyguard, unlock
keyguard, close app, lock keyguard, observe lockscreen - all in rapid
succession
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:09d67d2001c971285c0b73d0319a13b7272fe3f4)
Merged-In: I80313281219fd6b42fe3dffffde2ef90086b0e59
Change-Id: I80313281219fd6b42fe3dffffde2ef90086b0e59