Disallow bluetooth sharing by default for private profiles
This change adds the user restriction DISALLOW_BLUETOOTH_SHARING to the
list of default restrictions applied on private profile creation. This
will result in the "bluetooth" option not showing up in the "private"
tab of the sharesheet. Please note that users would still be able to use
the share via bluetooth option in the "primary" tab in sharesheet.

Test: atest UserManagerTest#testDefaultUserRestrictionsForPrivateProfile
Bug: 296357858
Change-Id: Idf07014a11690576d220d2599bb6a1f663a41789