Set the same density to all displays
It's better user experience to have the same density on all internal displays.

The density calculation uses the display size to come up with density values. We can use the size of one of the displays to calculate the values and then use the same values on the density slider for all displays. We use the size of the smallest display to make sure that the largest/smallest values aren't too extreme for the smallest display.

Bug: 324439718
Test: Change the density and see that the density has been applied to all displays.
Test: atest DisplaySizeDataTest
Test: atest ScreenResolutionFragmentTest
Flag: EXEMPT bugfix
Change-Id: Ie138c53b5042b78e3f32a133c5c1ba368c7b2b50

Former-commit-id: e60527feab7c83d11b8474603bc529add0659702