Consider transient-launch in waiting transitions
Following from [1], this also considers the waiting transitions when
checking for transient-launch activities.

This would trigger an issue when going from an app with the IME shown to
recents, where the transition to launcher was moved to waiting, meaning
launcher is no longer considered in transient launch. Because of this,
launcher was now passing WindowState#canBeImeTarget, which triggered a
change in the IME layering target from the app to launcher. This allows
InputMonitor#updateInputFocusRequest to request hiding the IME, making
it disappear from app view in recents.

  [1]: Icaa21a6be8a28bfa5369b954e645f72cd47c937c

Bug: 333175068
Test: atest ShowImeWhileEnteringOverviewTest
Change-Id: I4624c5e6f5da0e69c1adaa308555901cb2ec04c8

Former-commit-id: 80817dd4e5a4a14b0275dc5cd359f65d601bb68f