Remove redundant call of queryInputMethodServicesInternal (2nd try)
This is the 2nd attempt of submitting my previous CL [1], which got
reverted [2] due to a boottime regression (Bug 329414077).
We have already addressed the major contributor of the boot time
regression [3], so this CL should be ready to be submitted again.

To follow up my previous CL [4], which introduced
InputMethodSettingsRepository as a read-only mode again.

There was a redundant invocation of

  InputMethodManagerService#queryInputMethodServicesInternal,

in

  MyPackageMonitor#onFinishPackageChangesInternal().

This CL fixes it.

 [1]: I9437455cd9ad907a3587b8087c2aa2620ef4c8fd
      73b00964b2465f4eac7cdad67e6a66fac9dbccff
 [2]: Ib1ae7a2ac556799a7243ba2798bae3910a74d297
      10544cae5d0be68d562152cc32af66c832c83767
 [3]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      2029c969800226e5b6223da324409af279e05e46
 [4]: Ic0dd655fbd86b8ccce2b3298b4c70359a468f9ec
      4d16823a7bfc1a90f3d64f5b841c9c41a4cfeabf

Bug: 309837937
Bug: 333591545
Test: presubmit
Change-Id: I6c92bd92088aad022655d03acd6e535ed872d63b