Remove unnecessary query in IMMS#switchUserOnHandlerLocked() (2nd try)
This is the 2nd attempt of submitting my previous CL [1], which got
reverted [2] due to a boottime regression (Bug 329414077).
We have already addressed the major contributor of the boot time
regression [3], so this CL should be ready to be submitted again.

This is a follow up CL to my previous CL [4], which started using

  InputMethodSettingsRepository

as a on-memory data repository of InputMethodSettings to avoid
unnecessary queries to PackageManagerService again.

With this CL

  InputMethodManagerService#switchUserOnHandlerLocked()

also start using InputMethodSettingsRepository instead of querying
available IMEs for the new user dynamically.

This CL is supposed to improve the IME user switching latency.

 [1]: I9437455cd9ad907a3587b8087c2aa2620ef4c8fd
      61e91349a9ffb31307d88e34aa3df22a6572610c [formerly e5475c78d31a4dcb5b4e1ce2e12eaf30d0315555]
 [2]: Ib1ae7a2ac556799a7243ba2798bae3910a74d297
      ac63c7bb14dfe0f833a0aff7aa68f90a8d64dfc3 [formerly fa26cc98aa2eb8da2f3c615660e01946d65cbc4d]
 [3]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      d2667640bff477d140f6ba2a5a2b2643043ccefd [formerly bce89b5742781e680efac88cbc7179bf029638d9]
 [4]: Ic0dd655fbd86b8ccce2b3298b4c70359a468f9ec
      f8bcb12444053d7633347a09fd09c7ae68cc3470 [formerly 8130073e978ab0a826ecfab9bee1254551023a32]

Bug: 309837937
Bug: 329163064
Bug: 333591545
Test: presubmit
Change-Id: If1285c532d70496956938751519810ff008b0eda

Former-commit-id: 6171c37145448eab3db5e384beb3eb3cace6b3c6