Replace IMMS#mSettings with IMMS#mCurrentUserId (2nd try)
This is the 2nd attempt of submitting my previous CLs [1][2], which got
reverted due to a boottime regression (Bug 329414077) [3][4]. We have
already addressed the major contributor of the boot time regression [5],
so this CL should be ready to be submitted again.

This CL removes

  InputMethodManagerService#mSettings

as the same instance can be obtained from

  InputMethodSettingsRepository

as long as we keep track of the current user ID.

This is a mechanical refactoring CL. There must be no user observable
behavior change.

 [1]: Idae0fb6d1a9182efe917f81a1d881d488a1b57e5
      318189c511fec0149bdb0a1b91b2013eb578ed36 [formerly 745f9e77f457dcfaf09c70a4d1f8e3492bebf0dc]
 [2]: I4b7dc1b0e52e9bf25095eb2e3cc520d46b3869e5
      e0ae73ccc5028484c6bab126e738f9399f0af51f [formerly fe68163da287cd1966a73dd0a6982e9560f9538b]
 [3]: I0d65cb9edcdbdaea45b1fa80abfe917b72fb9a50
      07c8226d02015fb09f003b64be3183ab2cb4fdcb [formerly 4a81244af64286a76ff4f85d02ab2a85bf0ddc4d]
 [4]: I585e5c6745c40e51f1080b78367175990b2c5884
      57389ddfdc58d8cfbf16cec3f135582f6ae5d371 [formerly c579c41ae5d366887800f89192359d0e665fe182]
 [5]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      d2667640bff477d140f6ba2a5a2b2643043ccefd [formerly bce89b5742781e680efac88cbc7179bf029638d9]

Bug: 305849394
Fix: 329163064
Fix: 333591545
Test: presubmit
Change-Id: Iaa7f08c9081aac9abe0c95d522185632d575d6b8

Former-commit-id: dc1fd9b2ed220d1c3b7cf8881333a9d4d1044705