Allow work profile first run to include private space first run message
Reusing the work profile message template for now, iteration likely.

Forked the WorkProfileMessageController into ProfileMessageController,
generalized the logic a bit to support both Work and Private profiles.
Extracted the code that touched other parts of the system to make the
controller more testable. Aiming to keep the change to
MessageContainerController as minmimal as possible for now, with
basically all the changes behind the flag.

Bug: 324585817
Bug: 326941376
Test: atest ProfileMessageControllerTest MessageContainerControllerTest
Test: Manual testing with private/work/personal apps.
Flag: ACONFIG com.android.systemui.screenshot_private_profile TEAMFOOD

Change-Id: I80e0b9e72bbe8f11ea34d9f2e01dfc37769f7511