Let token animation notify removing starting window
AsyncRotationController may have executed all operations then it
becomes null, but the fade-in animation can still be running.
So if the specified type is animating, let the callback which]
is set by case Operation.ACTION_TOGGLE_IME in
AsyncRotationController#finishOp to call reportImeDrawnForOrganizer.

Otherwise if it reports ime drawn too early, the snapshot starting
window will be removed before the fade-in animation is completed,
which cause flickering.

Bug: 329781040
Test: ShowImeOnAppStartWhenLaunchingAppFromFixedOrientationTest
Change-Id: I23474f6e369ef390f46d5cdd95867e8a3070dbb7