feat(high contrast text): make high contrast text API public
This is so we can test it in CTS, so we can wait for the setting to take
effect before running the test.

It's also so Compose and other UI frameworks can check the setting and
do their own fixes and treatments of text.

Bug: 186567103
Bug: 326442372
Fix: 330241761
Flag: com.android.graphics.hwui.flags.high_contrast_text_small_text_rect
Test: atest cts/tests/tests/uirendering/src/android/uirendering/cts/testclasses/TextViewHighContrastTextTests.kt

Change-Id: I8dd5663e11c37df8aa86202d3a6f20a312ef69d5

Former-commit-id: ef60a7c0cccd0aa4d2f8999235f4ce16c9a5de9b