Fix bug with adding widgets via drag&drop when GH is empty
We recently removed the CTA tile in edit mode and we used to add a new
widget before the CTA card at the second to last index.
Fix is to update the index to insert to be the last in list.
Also cleaned up a few unused methods due to the removal of the CTA card.

Bug: b/334996701
Fixed: 334996701
Test: manual
Flag: ACONFIG com.android.systemui.communal_hub TEAMFOOD
Change-Id: I9fffd38fcbf62bf451a3033025fc99b81ac96d39