Fix a bug with the screen incorrectly turning off on power button press.
In certain situations, when the device is docked and the power button is
single pressed, the screen turns off instead of initiating a dream. It
appears that when this happens, DreamManagerService doesn't think the
device is charging even though PowerManagerService thinks it is. This
change reworks the charge detection logic in DreamManagerService so that
it works the same as the logic in PowerManagerService in an attempt to
fix the issue.

Bug: 329125239
Test: atest DreamManagerServiceTests
Flag: ACONFIG android.service.dreams.use_battery_changed_broadcast STAGING
Change-Id: I5344c12d471cbfc889bdf511a7e9e36c6c3b84e6