Override layout in display cutout mode when needed
This patch introduce a compat change id to support layout in display
cutout mode override. If the flag for the app is set to true, the app
will always layout in display cutout, regardless of the value it sets.

However, if the app is explicitly opt out the edge to edge behavior, we
respect the app's choice and ignore the override.

Test: Check table top behavior of video apps with the flag set
Bug: 332679525
Change-Id: I24e922d2200f0a3c362aeef36512c716590ce4d9

Former-commit-id: 88dba1b31f5f3641a13d58637156425626f7701f