Fix nullability in MediaController#getPlaybackInfo()
The playback info field is NonNull, so getPlaybackInfo() is artificially
returning a null value if the Binder transaction to system_server fails.
As per go/android-api-guidelines#framework-rethrow-remoteexception we should always be rethrowing
RemoteExceptions to provide a strong signal that the operation failed.

In practice, the returned value is never null, unless system_server has
died. Several CTS tests already assume this NonNull-ness.

Test: atest CtsMediaBetterTogetherTestCases
Bug: 333564788
Change-Id: Id18cce692d8d0b25486da427ef9860eb2f176aef

Former-commit-id: 5a3013b421d71e470c108eae082011809c7d6c3f