Create new Edit Mode
This creates the pipeline for the edit mode data as well as the
composable for InfiniteGridLayout. It also extracts common elements
between TileGrid and EditTileGrid.

Missing:
  * Finalized UI.
  * Drag and drop (and move action).
  * Logs, metrics

None of this code is flagged directly, but it's not connected to any
non-flagged code

Test: manual
Test: atest com.android.systemui.qs.panels
Flag: ACONFIG com.android.systemui.qs_ui_refactor DEVELOPMENT
Bug: 337021579

Change-Id: I351cc2ed2444b74478e80511bfd5cce2d14670d9

Former-commit-id: 2ccbc3c3738bebd476dcdc4d51de4cb3dcfe1074