Reapply "Finalize resources for VIC"
This reverts commit 40da5803ab3548d1e8983597b818d3f1ac736b91.

Instead of listing the resources twice, just mark the original ones as
removed. This will break any existing clients of the APIs until they
recompile, but given that flags weren't accesible to application for
most of this release (and still aren't available to most apps), this
should have minimal-to-no impact in practice.

Bug: 335304236
Test: atest --host  WearSdkRoboTests:WearSdkVersionTest
Change-Id: I7998a1f28e51e40e911ce6d51412232714993a84