Emulate Zen rule policy inheritance until edit UI is available
Before MODES_API, rules without a policy (or with a partially specified policy) would inherit the manual DND policy settings when calculating the consolidated policy. This change temporarily restores part of this behavior, because users have come to expect it (even though it's not applied consistently). In order to do so, whenever the manual policy is modified, we update the policies of rules that match either the default or previous manual policy.

This becomes obsolete with MODES_UI since the user can then view and edit those policies, and it will be clear that the NotificationPolicy only applies to manual DND (and this is also the reason that don't apply the new manual policy to rules that have their own policy pages today).

Flag: android.app.modes_api
Fixes: 337193321
Test: atest ZenModeHelperTest NotificationManagerZenTest
Change-Id: I2217b1f251608a1eea54072720b50b08eb3b4d39

Former-commit-id: 077d19070dc88a6f5ee391734b4988473d48f232