Placeholder for widgets pending install
This change adds a new subtype for communal widget content: pending.
This represents widgets that have been added in the hub but are being
installed. Now after a backup is restored, restored widgets whose
package is pending install will show up in the hub as a placeholder. The
UX is not final and may need polish.

Test: atest CommunalWidgetRepositoryImplTest
Test: atest CommmunalAppWidgetHostStartableTest
Test: atest CommunalInteractorTest
Test: manually with instructions at go/glanceable-hub-br
Bug: 330945453
Fix: 330945453
Flag: ACONFIG com.android.systemui.communal_hub TEAMFOOD
Change-Id: If03395e1ca22e2c93ab0ac41a240bbae301c1409

Former-commit-id: cfd183e2e8f46ff9e0bedde3aa9b4bd90091b6d8