Inline IMMS#setSelectedMethodIdLocked()
This is a follow up CL to our previous CL [1], which introduced
multiple utility methods into InputMethodManagerService with hoping
it'd help us encapsulate things.

Now that we are going to instantiate InputMethodBindingController in a
per-user basis, methods like

  @GuardedBy("ImfLock.class")
  private void setSelectedMethodIdLocked(
          @Nullable String selectedMethodId) {
      mBindingController.setSelectedMethodId(selectedMethodId);
  }

do not make much sense.

Let's inline this for better readability.

This is a mechanical change thus there must be no observable behavior
change.

 [1]: I9cbdd6606f7a44796f6074d0b2e904913bd48e51
      d759117c14e80e717950046d9aebc0637609046e

Bug: 325515685
Test: presubmit
Change-Id: I7d55ae9bd428d52dbb9664a79f42d6728a3148f2

Former-commit-id: c95c8f09f5acc494a394725fdafd9ff49136dfd4