Modify AppStartInfo logic for setting state
Now that first frame drawn timestamp is being reported, use this to determine
done state rather than onActivityFinished which is prior to first frame drawn.
This ensures that timestamps will all be added, as well as that the callback
will occur on a fully complete with timestmaps object.

Bug: 338125667
Test: manually check timestamps and state for various starts
Flag: android.app.app_start_info_timestamps
Change-Id: I889345bb1120273d1d6e6f86848dd12f789da721

Former-commit-id: 3f3932faa01e372adc4052ce4acf43a2cf6999c3