Add swipe gesture indicator for glanceable hub
This adds a handle to both the overlay and the Blank scene in the
communal SceneTransitionLayout. Long-term there will need to be a
better solution.

The handle in the Blank scene is not the correct color as it's
underneath the bottom scrim, which dims it slightly. Neither scrim
transitions super well or adjuts colors for underlying content.

See bug for video

Bug: 339667383
Test: atest DreamOverlayContainerViewControllerTest
Flag: com.android.systemui.glanceable_hub_gesture_handle
Change-Id: Ia5be42ff22fcbe3048f06e1a4937884f56232649