Prefer defining local variables over userData.mBindingController
UserData#mBindingController is a final field and is guaranteed to be
the same instance no matter when it gets read. Thus we should be able
to introduce a local variable to reference to it for better
readability.

This is a safe and mechanical code refactoring CL. There must be no
obsrevable behavior change.

Bug: 325515685
Test: presubmit
Change-Id: I6758a0ffcfa41665880facca54289cb81d6fd1f4

Former-commit-id: 755a38a185d96c0d71a5eee23fab194e317ca810