Simplify IMMS.MyPackageMonitor
With my previous changes [1][2], long and/or frequent operations in

 MyPackageMonitor#onFinishPackageChangesInternal()

are expected to be no longer contributing to UI janks. Now it is time
to focus more on the correctness rather than trying to optimize things
in a complicated way. Complicated optimizations make it difficult for
us to diagnose issues like Bug 335281466.

This CL basically replaces my past optimizations [3][4] in a more
strict and straightforward way. We first create a new InputMethodMap
based on the latest results returned from the PackageManagerService,
then see if there is any actual data change from the current
InputMethodMap. If those two InputMethodMap instances are exactly the
same, it means that no further update is necessary.

There should be no semantic observable change.

 [1]: I263cd49dd4d64b64136acc3dad469f83a862ce97
      8969f9ff34bdac07459a16d6a134e645f58f8eb2 [formerly db19711e043340805ea94548ed31012122d913a6]
 [2]: Icc84fe9d0d3b5842b4d5ee51b12b0d3cd10a1d24
      d7573a49d90a151d696873490221cf2bbe763cfb [formerly 81a9a1ba6647bf6698075db49c9667153eb6eb5b]
 [3]: I7b69c349318ce06a48d03a4468cf2c45bfb73dc2
      bfc7b2c467bdc8854cb34a6d3e3300173701496e
 [4]: I063688297156188f68fe0b55a46d72f2e811dc88
      2645504062852c240f4798216bf7c61b4923c3df

Bug: 329703038
Test: presubmit
Change-Id: I6e9de6ff5c6dd55e3c40f23586110cb713c62851

Former-commit-id: 6bdc56c6962dbaa5c9576b6d1d50de95a18712ff