Update the leash position when gaining a new control
If there is an animation playing, the leash position will be updated
after calling InsetsAnimationControlRunner#updateSurfacePosition in the
next animation frame or the end of animation.

However, if there is no animation playing. The client wouldn't update
the leash position if the surface position of the insets source window
is changed at the server side. This is fine in most of time because the
server will update the leash position. But if the client is playing the
animation while the server is updating the leash position, and then the
client finishes the animation right before receiving the new control,
the position can be overwritten by the client animation.

This CL updates the leash position when the client gains a new control
if there is no leash-animation playing, like we did to the visibility.

Bug: 337018949
Flag: NA
Test: Turn screen off while the device is unfolded, fold the device,
      and then turn screen on. See if the taskbar position is expected.
Merged-In: I5ba0c33ca4542d27bb4dee4ce849f34721f20b14
Change-Id: I5ba0c33ca4542d27bb4dee4ce849f34721f20b14
(cherry picked from commit 6de340a93e14bebb46a0bd97de371b51b1e3601e)