Apply autobrightness adjustments while checking the strategy validity
For checking the validity of autobrightness, we first check check if
the value in autobrightness controller is a valid one, and qualify the
strategy accordingly. However, the adjustments are not applied when this
validation is done, but instead applied when the actual brightness is
loaded while executing the strategy. While this is fine for most cases,
for some corner cases(like brightness being invalid because of invalid
lux), auto brightness adjustments are not applied.

Bug: 341643213
Test: atest AutomaticBrightnessStrategyTest
Change-Id: I4ebce5905b9237af8bb747158a3a9fcc565617f4

Former-commit-id: 786c3514be5d95806674db0af2b9242a9ca4b6e4