Conditionally initialize IMMS#mCurrentUserId with a main user
With an assumption that we would enable concurrent multi-user IME
support only in an environment where user 0 is not the main user, this
CL changes how

  InputMethodManagerService#mCurrentUserId

is initialized when IMMS#mExperimentalConcurrentMultiUserModeEnabled
is set to true. With this CL it will be initialized to be the first
found main user ID instead of the current user ID when IMMS is
initialized.

Note that this behavior change is fully guarded behind

  IMMS#mExperimentalConcurrentMultiUserModeEnabled

hence there must be no observable behavior changes in other
environments.

Bug: 341558132
Flag: android.view.inputmethod.concurrent_input_methods
Test: presubmit
Test: atest FrameworksInputMethodSystemServerTests
Change-Id: Ie2ee161bea771c5928b504bee494813ac8e1fb7b

Former-commit-id: cd4a8a114cdf99cf747c9b5f10c77163448f7ef0