Handle failed AppStartInfo record write
The IOException this catches can leave the write in a bad state causing it to fail later and crash system server.
Store whether individual record writes succeeded and then fail the write if not.
Also manually delete record as it's unlikely to work any better on the next persist.

This was manually confirmed to be the problem by throwing IOException in AppStartInfo#writeToProto and confirming system server crashes.

Test: manually throw an IOException in AppStartInfo#writeToProto, ensure write fails and system server lives
Bug: 342319837
Flag: EXEMPT - bugfix
Change-Id: Id0fa47fdade53bd8f262fbdb8aa1f75136aadcbd

Former-commit-id: 4b99fe8aec29fe67ccb753cad382218f95fe75f5