Move changeScene calls to the application scope
SceneTransitionLayout now enforces that state changes are done on the
same thread the STL was made on, which in our case is just the main
thread.

Move our calls to the main thread to fix timeout and potentially other
places where we go to/from the glanceable hub from breaking.

Bug: 342470602
Fix: 342470602
Test: manually verified that hub times out successfully
Flag: com.android.systemui.glanceable_hub
Change-Id: I980a314f3198f04611a0e9e979326ba0d2db34e7

Former-commit-id: 3cb501f26f5188ef0617d0ba3a749d31cfa6461f