Refactor communal scene logic to standalone interactor
This moves the logic out of CommunalInteractor and into a standalone
interactor. This simplifies CommunalInteractor, and allows for the new
interactor to be injected into code which would otherwise cause
dependency cycles.

This change doesn't move all existing tests. This will be done in a
follow-up change to move the unit tests to a separate file.

This change does have one minor change - we remove the communalAvailable
check when calculating the scene. This is safe because the scene now
comes directly from STL. So if communal isn't available, its guaranteed
to be reset back to Blank.

Test: atest SystemUiRoboTests
Flag: com.android.systemui.glanceable_hub
Bug: 331591373
Change-Id: I5be1d0483732ff0cbabfddd2095f80c58e42c6a5

Former-commit-id: 0f4134713cc63566414c867d20f6d0cddf5a9409