Offload Doze Override should trigger suspend blocker
Unlike most display-state-changes, Offload state changes are not
triggered by a PowerRequest.  Because of this PowerManager isn't aware
of the change and doesn't attempt to keep the device awake.  As a
result, offload transitions sometimes went into suspend before they were
complete, preventing offload from running when it should.

This change adds a suspend blocker to offload doze override requests,
and notifies PowerManager when it's processing a change.

Test: atest com.android.server.display
Bug: 336599466
Change-Id: Ib9a6eeeeb3bc0c2c66ac02dff1aec61203e6b04c

Former-commit-id: a8c72597f8e12b7ad1b97376800837a3b7992dd7