Move media output state logic to the interactor.
Apparently, there is a defference in active sessions returned by the
MediaSessionManager when starting a call and getting one. This caused
the panel to correctly determine the device when starting a call, but
not when getting one.

Now Media Output Component state is completely detached from the media
session when there is an ongoing call.

Flag: com.android.systemui.new_volume_panel
Fixes: 342976548
Test: manual on the phone
Test: atest MediaOutputComponentInteractorTest
Test: atest MediaOutputViewModelTest
Change-Id: Ibeda4a7ac10053314e6e436fc5a458429d9293d7

Former-commit-id: 4752ce9bc8e10ac17517623c90ee6716cf52ebda