Force Now Playing to re-layout after lock icon moves when folding
With DeviceEntryUdfpsRefactor ON and KeyguardBottomAreaRefactor OFF,
now playing is not able to constrain off of the lock icon at the correct
time, which allowed for the possibility of layout timing issue.

Once the constraints are set on the lock icon, we will now re-layout the
ambient indication container (now playing).

This temporary solution will be removed once the
KeyguardBottomAreaRefactor flag is enabled.

Fixes: 340566200
Test: manual - activate now playing on foldable, unfold and re-fold
Flag: com.android.systemui.device_entry_udfps_refactor
Change-Id: I22e238843dd62ff0281c9b73ff1accb7065fe5a6