InputManager: fix grammar in verifyInputEvent JavaDoc
It seems that adding @NonNull to a parameter causes "This value cannot
be null" to be appended to its JavaDoc on developer.android.com. Because
the @param tag for the event parameter didn't end with a full stop, this
caused the JavaDoc to read "The InputEvent to check This value cannot
be null".

The @return tag was also missing a full stop.

Bug: 245989146
Change-Id: I015cf07cb94640b585242a485d4d4c44fc739255
Test: none
Flag: DOCS_ONLY